feat: Secutity and handler exceptions, not tested yet

+++ src/main/java/com/acme/backendimagia/application/artwork_recognition/ArtworkService.java
@@ -5,6 +5,7 @@ import com.acme.backendimagia.domain.model.Artwork;
 import com.acme.backendimagia.domain.model.User;
 import com.acme.backendimagia.domain.repository.ArtworkRepository;
 import com.acme.backendimagia.domain.repository.UserRepository;
+import com.acme.backendimagia.shared.exception.BusinessExceptions;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -22,6 +23,11 @@ public class ArtworkService {
     private UserRepository userRepository;
 
     public List<ArtworkDTO> getSavedArtworksByUser(Long userId) {
+        // Verificar si el usuario existe
+        if (!userRepository.existsById(userId)) {
+            throw new BusinessExceptions.ResourceNotFoundException("Usuario no encontrado con ID: " + userId);
+        }
+
         List<Artwork> artworks = artworkRepository.findByUserId(userId);
         return (artworks != null ? artworks : List.<Artwork>of()).stream()
                 .map(this::mapToDTO)
@@ -36,14 +42,15 @@ public class ArtworkService {
     }
 
     public ArtworkDTO getArtworkByName(String name) {
-        Optional<Artwork> obraOptional = artworkRepository.findFirstByName(name);
-        return obraOptional.map(this::mapToDTO).orElse(null);
+        return artworkRepository.findFirstByName(name)
+                .map(this::mapToDTO)
+                .orElseThrow(() -> new BusinessExceptions.ResourceNotFoundException("Obra de arte no encontrada con nombre: " + name));
     }
 
     @Transactional
     public void createArtwork(ArtworkDTO artworkDTO) {
         User user = userRepository.findById(artworkDTO.getUserId())
-                .orElseThrow(() -> new RuntimeException("Usuario no encontrado con ID: " + artworkDTO.getUserId()));
+                .orElseThrow(() -> new BusinessExceptions.ResourceNotFoundException("Usuario no encontrado con ID: " + artworkDTO.getUserId()));
 
         Artwork obra = new Artwork();
         obra.setUser(user);
@@ -57,12 +64,17 @@ public class ArtworkService {
     }
 
     public void deleteArtwork(Long obraId) {
+        if (!artworkRepository.existsById(obraId)) {
+            throw new BusinessExceptions.ResourceNotFoundException("Obra de arte no encontrada con ID: " + obraId);
+        }
         artworkRepository.deleteById(obraId);
     }
 
+
     public ArtworkDTO getArtworkById(Long obraId){
         Optional<Artwork> obraOptional = artworkRepository.findById(obraId);
-        return obraOptional.map(this::mapToDTO).orElse(null);
+        return obraOptional.map(this::mapToDTO)
+                .orElseThrow(() -> new BusinessExceptions.ResourceNotFoundException("Obra de arte no encontrada con ID: " + obraId));
     }
 
     // Método privado para mapear ObraDeArte a ObraDeArteDTO
